requirements.txt
adal==1.2.2
applicationinsights==0.11.9
arrow==0.15.4
asn1crypto==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.24
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.48
azure-eventgrid==1.3.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage==0.36.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
azureml-automl-core==1.0.83.1
azureml-contrib-pipeline-steps==1.0.83
azureml-core==1.0.83
azureml-dataprep==1.1.35
azureml-dataprep-native==13.2.0
azureml-pipeline==1.0.83
azureml-pipeline-core==1.0.83
azureml-pipeline-steps==1.0.83
azureml-sdk==1.0.83
azureml-telemetry==1.0.83
azureml-train==1.0.83
azureml-train-automl-client==1.0.83
azureml-train-core==1.0.83
azureml-train-restclients-hyperdrive==1.0.83
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
contextlib2==0.6.0.post1
coverage==4.5.4
cryptography==2.8
dataclasses-json==0.3.6
dateparser==0.7.2
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
docker==4.1.0
dotnetcore2==2.1.11
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
fpdf==1.7.2
fusepy==3.0.1
fuzzywuzzy==0.17.0
idna==2.8
importlib-metadata==1.5.0
inflect==4.1.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
jaraco.itertools==5.0.0
jedi==0.16.0
jeepney==0.4.2
Jinja2==2.11.1
jmespath==0.9.4
jsonpickle==1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.1
marshmallow==3.3.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mistune==0.8.4
moment==0.8.2
more-itertools==8.2.0
msrest==0.6.10
msrestazure==0.6.2
mypy==0.750
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
ndg-httpsclient==0.5.1
notebook==6.0.3
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
price-parser==0.3.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pyarrow==0.15.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyPDF2==1.26.0
pyrsistent==0.15.6
PySocks==1.7.1
pyspark==2.4.4
pytesseract==0.3.1
pytest==5.3.0
pytest-asyncio==0.10.0
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.5
python-Levenshtein==0.12.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
regex==2019.11.1
requests==2.22.0
requests-mock==1.7.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.89
SecretStorage==3.1.2
Send2Trash==1.5.0
six==1.14.0
spacy==2.2.3
stringcase==1.2.0
terminado==0.8.3
testpath==0.4.4
times==0.7
toml==0.10.0
tornado==6.0.3
tox==3.14.2
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
typing-inspect==0.5.0
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0